+++ launch_env.sh
@@ -12,5 +12,5 @@ fi
 
 export STAGING_ROOT="/data/safe_staging"
 
-export FINGERPRINT="CHANGAN_Z6"
-export SKIP_FW_QUERY="1"
+#export FINGERPRINT="CHANGAN_Z6"
+#export SKIP_FW_QUERY="1"
