+++ opendbc_repo/opendbc/dbc/changan_z6_pt.dbc
@@ -44,6 +44,9 @@ BO_ 368 GW_170: 8 ESP
  SG_ COUNTER : 51|4@0+ (1,0) [0|15] "" XXX
  SG_ CHECKSUM : 63|8@0+ (1,0) [0|255] "" XXX
 
+BO_ 378 VEHICLE_SPEED: 8 XXX
+ SG_ VEHICLE_SPEED : 39|16@0+ (0.05,0) [0|250] "km/h" XXX
+
 BO_ 382 GW_17E: 8 XXX
  SG_ EPS_MeasuredTorsionBarTorque : 7|12@0+ (0.01,0) [0|4095] "" XXX
  SG_ EPS_RollingCounter_17E : 51|4@0+ (1,0) [0|15] "" XXX
@@ -68,12 +71,16 @@ BO_ 392 NEW_MSG_188: 8 XXX
 
 BO_ 406 GW_196: 8 XXX
  SG_ EMS_RealAccPedal : 20|1@0+ (1,0) [0|1] "" XXX
+ SG_ BRAKE_PRESSED : 0|1@0+ (1,0) [0|1] "" XXX
+ SG_ GAS_PEDAL_USER : 15|8@0+ (1,0) [0|0] "" XXX
  SG_ COUNTER : 51|4@0+ (1,0) [0|15] "" XXX
  SG_ EMS_BrakePedalStatus : 54|1@0+ (1,0) [0|1] "" XXX
  SG_ CHECKSUM : 63|8@0+ (1,0) [0|255] "" XXX
 
 BO_ 422 NEW_MSG_1A6: 8 XXX
  SG_ NEW_SIGNAL_1 : 7|16@0+ (0.78125,0) [0|65535] "" XXX
+ SG_ BRAKE_PRESSED : 0|1@0+ (1,0) [0|1] "" XXX
+ SG_ BRAKE : 31|8@0+ (1,0) [0|255] "" XXX
 
 BO_ 438 NEW_MSG_1B6: 8 XXX
  SG_ NEW_SIGNAL_1 : 7|16@0+ (1,0) [0|65535] "" XXX
